+++ attribution/pkg/impression/kv/aerospike/aerospike.go
@@ -5,7 +5,7 @@ import (
     "strconv"
     "time"
 
-    "github.com/TencentAd/attribution/attribution/pkg/impression/kv"
+    "github.com/TencentAd/attribution/attribution/pkg/impression/kv/opt"
     "github.com/aerospike/aerospike-client-go"
 )
 
@@ -15,11 +15,11 @@ var (
 
 type Aerospike struct {
     client *aerospike.Client
-    option *kv.Option
+    option *opt.Option
 }
 
 func (a *Aerospike) Has(key string) (bool, error) {
-    k, _ := aerospike.NewKey(Namespace, kv.Prefix, key)
+    k, _ := aerospike.NewKey(Namespace, opt.Prefix, key)
     r, err := a.client.Get(nil, k)
     if r == nil || err != err {
         return false, err
@@ -28,12 +28,12 @@ func (a *Aerospike) Has(key string) (bool, error) {
 }
 
 func (a *Aerospike) Set(key string) error {
-    k, _ := aerospike.NewKey(Namespace, kv.Prefix, key)
+    k, _ := aerospike.NewKey(Namespace, opt.Prefix, key)
     aerospike.NewWritePolicy(0, uint32(a.option.Expiration / time.Second))
     return a.client.Put(nil, k, nil)
 }
 
-func New(option *kv.Option) (*Aerospike, error) {
+func New(option *opt.Option) (*Aerospike, error) {
     ip, port1, _ := net.SplitHostPort(option.Address)
     port, _ := strconv.Atoi(port1)
     client, err := aerospike.NewClient(ip, port)
