+++ attribution/pkg/impression/kv/redis/redis.go
@@ -4,7 +4,7 @@ import (
     "context"
     "time"
 
-    "github.com/TencentAd/attribution/attribution/pkg/impression/kv"
+    "github.com/TencentAd/attribution/attribution/pkg/impression/kv/opt"
     "github.com/go-redis/redis/v8"
 )
 
@@ -14,12 +14,12 @@ var (
 
 type Redis struct {
     client *redis.Client
-    option *kv.Option
+    option *opt.Option
 }
 
 func (r *Redis) Has(key string) (bool, error) {
     ctx, _ := context.WithTimeout(context.Background(), DefaultClientTimeOut)
-    _, err := r.client.Get(ctx, kv.Prefix + key).Result()
+    _, err := r.client.Get(ctx, opt.Prefix + key).Result()
     if err == redis.Nil {
         return false, nil
     }
@@ -33,10 +33,10 @@ func (r *Redis) Has(key string) (bool, error) {
 
 func (r *Redis) Set(key string) error {
     ctx, _ := context.WithTimeout(context.Background(), DefaultClientTimeOut)
-    return r.client.Set(ctx, kv.Prefix + key, nil, r.option.Expiration).Err()
+    return r.client.Set(ctx, opt.Prefix + key, nil, r.option.Expiration).Err()
 }
 
-func New(option *kv.Option) (*Redis, error) {
+func New(option *opt.Option) (*Redis, error) {
     client := redis.NewClient(&redis.Options{
         Addr:     option.Address,
         Password: option.Password,
